IPAD-00088 GymRoom modularization — PeerMirror package + GymRoomDisplay target IPAD-00088A PeerMirror package extraction
