+++ SharedModels/Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version: 6.1
+// swift-tools-version: 6.2
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -9,7 +9,8 @@ let package = Package(
     platforms: [
         .iOS("26.0"),
         .watchOS("26.0"),
-        .macOS("26.0")
+        .macOS("26.0"),
+        .tvOS("26.0")
     ],
     products: [
         .library(
